+++ test/matching_bytes.h
@@ -12,5 +12,5 @@
  */
 class ProtocolMatchingBytes: public Protocol {
 public:
-  PacketStatus isPacket(const RS485BusBase& bus, const int startIndex, int& endIndex) const;
+  PacketStatus isPacket(const RS485BusBase& bus, const size_t startIndex, size_t& endIndex) const;
 };
